+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## v1.1 - 2017-12-06
+### Changed
+- TOC version for 7.3
+### Added
+- Font size control
+- Reset for font size and position
+- Artifact weapon icon next to the AP text
+- Artifact Power left to get the next rank (on tooltip)
+- Dynamic color (from red to green) on the complete AP %
+- Thousand (K), Million (M) and Billion (B) to the AP values
+- Points spent on the current artifact weapon (on tooltip)
+### Removed
+- Number separators as they're no longer needed
+- Artifact Knowledge level since everyone has it at max level 55
+- Stamina from points spent (was pretty useless and never well implemented)
+### Fixed
+- Major problem causing the game to completly freeze
+
 ## v1.0 - 2017-04-02
 ### Release
 - Initial Release